+++ django_honeyguard/__init__.py
@@ -1,6 +1,6 @@
 """django-honeyguard package metadata and default app config."""
 
-__version__ = "0.1.3"
+__version__ = "1.0.0"
 
 default_app_config = "django_honeyguard.apps.HoneyGuardConfig"